support std::optional or std::experimental::optional
Newer compilers, specifically the latest Clang on OSX, have removed
`std::experimental::optional` entirely. This leads to two issues:

1) The imports and namespaces for `optional` have changed.
2) `optional` is only supported for C++17 and above, not C++14.

This commit uses CMake to detect which namespace contains optional,
sets a macro with the detected import and namespace, and updates all
usages of optional to use these detected import and namespaces.

Addditionally, if optional has been moved out of the experimental
namespace, it upgrades the required standard to C++17.