boost: upgrade to Boost 1.71
The ASIO `executor.get_io_context()` dprecated APIs were removed in
Boost 1.70.  This commit updates our code to use the new
executor-based APIs and sets the minimum Boost version 1.70.

This upgrade is necessary to build with the latest compilers on OSX.
Boost 1.66 (the prior version) does not compile on the latest Clang.